Create Fibonacci_Parallel.cpp
This toy example uses a task graph to recursively compute Fibonacci numbers in parallel. Results of the threads are generated via references to the spawning threads. A threshold (in terms of Fibonacci number to be generated) is used in order to not generate too many threads.

Be aware that creation of threads and the synchronisation costs time.

Examples:

too little work-load for the threads n=10, threshold=2
too many threads n=20, threshold=2
reasonable workload n=40, threshold=35